Move context to new files and export for use downstream
