+++ src/components/DatasetTableTab/ManageColumnsContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const ManageColumnsContext = createContext({})
+
+export default ManageColumnsContext;
